% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_equivalence.R
\name{brier_equivalence}
\alias{brier_equivalence}
\title{brier_equivalence function}
\usage{
brier_equivalence(m_a, m_b, alpha = 0.05, delta = 1)
}
\arguments{
\item{m_a}{logistic regression model \eqn{M_A}}

\item{m_b}{logistic regression model \eqn{M_B}}

\item{alpha}{significance level \eqn{\alpha} (defaults to 0.05)}

\item{delta}{equivalence sensitivity level \eqn{\delta_B} (defaults to 1)}
}
\value{
the \emph{AB} notation refers to the fit of model A to the data of
     model B (and \emph{BA} for vice versa) \describe{
  \item{\code{bs_mat}}{Brier scores matrix}
  \item{\code{bs_ab_equiv}}{Are models \eqn{M_A,M_B} producing equivalent
      Brier scores for \eqn{M_B}'s training data? (boolean)}
  \item{\code{bs_ba_equiv}}{Are models \eqn{M_A,M_B} producing equivalent
      Brier scores for \eqn{M_A}'s training data? (boolean)}
  \item{\code{test_stat_ab_l}}{\eqn{t_L} for the AB test}
  \item{\code{test_stat_ab_u}}{\eqn{t_U} for the AB test}
  \item{\code{test_stat_ba_l}}{\eqn{t_L} for the BA test}
  \item{\code{test_stat_ba_u}}{\eqn{t_U} for the BA test}
  \item{\code{crit_ab}}{level-\eqn{\alpha} critical value the AB test}
  \item{\code{crit_ba}}{level-\eqn{\alpha} critical value the BA test}

}
}
\description{
This function takes two logistic regression models \eqn{M_A, M_B},
    significance level \eqn{\alpha} and sensitivity level \eqn{\delta_B}.
    It checks whether the models perform equivalently on each of their
    training datasets and returns various figures.
}
\keyword{brier}
\keyword{equivalence}
