% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrierScore.R
\name{BrierScore}
\alias{BrierScore}
\title{BrierScore function}
\usage{
BrierScore(pi, y)
}
\arguments{
\item{pi}{the predictions vector}

\item{y}{the obsrevations vector}
}
\description{
This function takes a predictions vector \eqn{\pi} and matching obsrevations \eqn{y} and returns the Brier score for the predictions.
}
\examples{
pi_vec <- runif(10)
y_vec <- rbinom(10, 1, pi_vec)
BrierScore(pi_vec, y_vec)
}
\keyword{Brier}
