% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_vector_equivalence.R
\name{coef_vector_equivalence}
\alias{coef_vector_equivalence}
\title{coef_vector_equivalence function}
\usage{
coef_vector_equivalence(data_a, data_b, formula, delta, alpha = 0.05)
}
\arguments{
\item{data_a}{logistic regression model \eqn{M_A}}

\item{data_b}{logistic regression model \eqn{M_B}}

\item{formula}{logistic regression formula}

\item{delta}{equivalence sensitivity level \eqn{\delta_\beta}}

\item{alpha}{significance level \eqn{\alpha} (defaults to 0.05)}
}
\value{
\describe{
  \item{\code{critical value}}{level-\eqn{\alpha} critical value}
}
}
\description{
This function takes two logistic regression models \eqn{M_A, M_B},
    significance level \eqn{\alpha} and sensitivity level \eqn{\delta_B}.
    It checks whether the coefficient vectors are equivalent.
}
\keyword{beta}
\keyword{equivalence}
\keyword{equivalence,}
\keyword{vector}
