% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/le_base_data.R
\name{le_base_data}
\alias{le_base_data}
\title{le_base_data function}
\usage{
le_base_data(m_a, m_b)
}
\arguments{
\item{m_a}{a logistic regression model}

\item{m_b}{a logistic regression model}
}
\description{
This function takes two logistic regression models \eqn{M_A, M_B} and
     arranges their data for the equivalence functions.
}
