% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptg_stud_data.R
\docType{data}
\name{ptg_stud_data}
\alias{ptg_stud_data}
\title{Student Performance Data Set}
\format{
An object of class \code{data.frame} with 649 rows and 31 columns.
}
\source{
\href{https://archive.ics.uci.edu/ml/datasets/student+performance}
}
\usage{
"ptg_stud_f_train"

"ptg_stud_f_test"

"ptg_stud_m_train"

"ptg_stud_m_test"

Data from a student achievement in secondary education of two Portuguese
schools. Full attribute description could be found in the source webpage.

The original Portuguese language acheivement data was split by gender (F/M)
\eqn{n_f=383, n_m=266}. The target variable \code{G3} was converted to
binary, \code{final_fail} which indicates the cases where \code{G3 < 10}.
Next, the data was split into train and test datasets.
}
\description{
Student Performance Data Set
}
\references{
P. Cortez and A. Silva. Using Data Mining to Predict Secondary
School Student Performance. In A. Brito and J. Teixeira Eds., Proceedings
of 5th FUture BUsiness TEChnology Conference (FUBUTEC 2008) pp. 5-12,
Porto, Portugal, April, 2008, EUROSIS, ISBN 978-9077381-39-7.
}
\seealso{
\href{http://www3.dsi.uminho.pt/pcortez/student.pdf}
}
\keyword{datasets}
