% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptg_stud_data.R
\docType{data}
\name{ptg_stud_f_test}
\alias{ptg_stud_f_test}
\title{Student Performance Data Set - female testing data}
\format{
An object of class \code{data.frame} with 77 rows and 30 columns.
}
\usage{
ptg_stud_f_test
}
\description{
Student Performance Data Set - female testing data
}
\keyword{datasets}
