% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptg_stud_data.R
\docType{data}
\name{ptg_stud_f_train}
\alias{ptg_stud_f_train}
\title{Student Performance Data Set - female training data}
\format{
An object of class \code{data.frame} with 306 rows and 30 columns.
}
\usage{
ptg_stud_f_train
}
\description{
Student Performance Data Set - female training data
}
\keyword{datasets}
