% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptg_stud_data.R
\docType{data}
\name{ptg_stud_m_test}
\alias{ptg_stud_m_test}
\title{Student Performance Data Set - male testing data}
\format{
An object of class \code{data.frame} with 53 rows and 30 columns.
}
\usage{
ptg_stud_m_test
}
\description{
Student Performance Data Set - male testing data
}
\seealso{
\code{ptg_stud_data}
}
\keyword{datasets}
