% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function}
\usage{
sigmoid(theta)
}
\arguments{
\item{theta}{the linear predictor}
}
\description{
This function takes a number \eqn{\theta\in\mathbb{R}} and returns its respective probability in \eqn{[0,1]}.
}
\examples{
sigmoid(0)
}
\keyword{sigmoid}
