% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_equivalence.R
\name{theta_equivalence}
\alias{theta_equivalence}
\title{coef_vector_equivalence function}
\usage{
theta_equivalence(m_a, m_b, alpha = 0.05, delta = 1, r = 0)
}
\arguments{
\item{m_a}{logistic regression model \eqn{M_A}}

\item{m_b}{logistic regression model \eqn{M_B}}

\item{alpha}{significance level \eqn{\alpha} (defaults to 0.05)}

\item{delta}{equivalence sensitivity level \eqn{\delta_\beta} (defaults to 1)}

\item{r}{ratio of allowed 'flips' (defaults to 0)}
}
\value{
the \emph{AB} notation refers to the fit of model A to the data of
     model B (and \emph{BA} for vice versa) \describe{
  \item{\code{ab_equiv}}{Are models \eqn{M_A,M_B} producing equivalent
      log-odds for \eqn{M_B}'s training data? (boolean)}
  \item{\code{ba_equiv}}{Are models \eqn{M_A,M_B} producing equivalent
      log-odds for \eqn{M_A}'s training data? (boolean)}
  \item{\code{test_stat_ab}}{The AB test statistic}
  \item{\code{test_stat_ba}}{The BA test statistic}
  \item{\code{crit_ab}}{level-\eqn{\alpha} critical value the AB test}
  \item{\code{crit_ba}}{level-\eqn{\alpha} critical value the BA test}
  \item{\code{xi_bar_ab}}{Mean \eqn{\xi} value for the AB test}
  \item{\code{xi_bar_ba}}{Mean \eqn{\xi} value for the BA test}
  \item{\code{delta_ab}}{Sensitivity level for the AB equivalence test
      (useful in case of indicating \code{r} value)}
  \item{\code{delta_ba}}{Sensitivity level for the BA equivalence test
      (useful in case of indicating \code{r} value)}
  \item{\code{p_val_ab}}{P-value for the AB test statistic}
  \item{\code{p_val_ba}}{P-value for the BA test statistic}
}
}
\description{
This function takes two logistic regression models \eqn{M_A, M_B},
    significance level \eqn{\alpha} and sensitivity level \eqn{\delta_B}.
    It checks whether the log-odds vectors are equivalent.
}
\keyword{beta}
\keyword{equivalence}
\keyword{equivalence,}
\keyword{vector}
