% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_vector_equivalence.R
\name{coef_vector_equivalence}
\alias{coef_vector_equivalence}
\title{coef_vector_equivalence function}
\usage{
coef_vector_equivalence(m_a, m_b, alpha = 0.05, delta = 1)
}
\arguments{
\item{m_a}{logistic regression model \eqn{M_A}}

\item{m_b}{logistic regression model \eqn{M_B}}

\item{alpha}{significance level \eqn{\alpha} (defaults to 0.05)}

\item{delta}{equivalence sensitivity level \eqn{\delta_\beta} (defaults to 1)}
}
\value{
\describe{
  \item{\code{test stat}}{Equivalence test statistic}
  \item{\code{critical value}}{level-\eqn{\alpha} critical value}
  \item{\code{ncp}}{non-centrality parameter}
  \item{\code{p_val}}{P-value for the test statistic}
  \item{\code{df}}{degrees of freedom}
  \item{\code{equivalence}}{are the coefficient vectors equivalent?
       (boolean)}
}
}
\description{
This function takes two logistic regression models \eqn{M_A, M_B},
    significance level \eqn{\alpha} and sensitivity level \eqn{\delta_B}.
    It checks whether the coefficient vectors are equivalent.
}
\keyword{beta}
\keyword{equivalence}
\keyword{equivalence,}
\keyword{vector}
