% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function}
\usage{
sigmoid(theta)
}
\arguments{
\item{theta}{the linear predictor}
}
\value{
the sigmoid probability
}
\description{
This function takes a number \eqn{\theta} and returns its
    respective sigmoid probability \eqn{\frac{e^{theta}}{1+e^{theta}}}.
    This is used in logistic regression to model \eqn{P(y=1|x)}.
}
\examples{
sigmoid(0)
}
